Removed test endpoints from squareController, renamed project-standards.md
