+++ client/src/components/squareGame/SquareGame.tsx
