+++ client/src/components/squareGame/partials/square/Square.tsx
