+++ server/api/Controllers/SquareController.cs
@@ -42,17 +42,6 @@ public class SquareController(ISquareRepository squareRepository) : ControllerBa
 
         }
     }
-    [HttpGet("test-409")]
-    public ActionResult Test409Error()
-    {
-        ModelState.AddModelError("Test conflict", "Model state test conflict.");
-        return Conflict(new { type = "Conflict", title = "This color is used by previous square", status = 409, errors = ModelState });
-    }
-    [HttpGet("test-400")]
-    public ActionResult Test400Error()
-    {
-        ModelState.AddModelError("Test badRequest", "Model state test badRequest.");
-        return BadRequest(ModelState);
-    }
+
 }
